+++ src/Beta.Famicom/Boards/Konami/VRC1.cs
@@ -19,7 +19,7 @@ namespace Beta.Famicom.Boards.Konami
             this.prg_page[3] = 0x1e000;
         }
 
-        public void R2A03Read(ushort address, ref byte data)
+        public void R2A03Read(int address, ref byte data)
         {
             if ((address & 0x8000) == 0x8000)
             {
@@ -27,7 +27,7 @@ namespace Beta.Famicom.Boards.Konami
             }
         }
 
-        public void R2A03Write(ushort address, byte data)
+        public void R2A03Write(int address, byte data)
         {
             switch (address & 0xf000)
             {
@@ -61,7 +61,7 @@ namespace Beta.Famicom.Boards.Konami
             throw new CompilerPleasingException();
         }
 
-        public void R2C02Read(ushort address, ref byte data)
+        public void R2C02Read(int address, ref byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -69,7 +69,7 @@ namespace Beta.Famicom.Boards.Konami
             }
         }
 
-        public void R2C02Write(ushort address, byte data)
+        public void R2C02Write(int address, byte data)
         {
             if ((address & 0x2000) == 0x0000)
             {
@@ -88,7 +88,7 @@ namespace Beta.Famicom.Boards.Konami
             throw new CompilerPleasingException();
         }
 
-        public bool VRAM(ushort address, out int a10)
+        public bool VRAM(int address, out int a10)
         {
             var x = (address >> 10) & 1;
             var y = (address >> 11) & 1;
